separate config into own file
